--- src/main/java/com/example/TechnicalAnalysis/Services/ConnectorService/Retriever.java
@@ -1,4 +1,0 @@
-package com.example.TechnicalAnalysis.Services.ConnectorService;
-
-public class Retriever {
-}
