--- src/main/java/com/example/TechnicalAnalysis/Services/DatabaseService/DatabaseReader.java
@@ -1,6 +1,0 @@
-package com.example.TechnicalAnalysis.Services.DatabaseService;
-
-public class DatabaseReader {
-    public void get() {
-    }
-}
