--- src/main/java/com/example/TechnicalAnalysis/Services/GitHubService/ActionsListener.java
@@ -1,4 +1,0 @@
-package com.example.TechnicalAnalysis.Services.GitHubService;
-
-public class ActionsListener {
-}
