--- src/main/java/com/example/TechnicalAnalysis/Services/GitHubService/Endpoints/EndpointsUtils/MapKeys.java
@@ -1,6 +1,0 @@
-package com.example.TechnicalAnalysis.Services.GitHubService.Endpoints.EndpointsUtils;
-
-public enum MapKeys {
-    COLLABORATORS,
-    COMMITS
-}
